refactor: Rename `Window.js` to `Window‑impl.js`
