+++ lib/api.js
@@ -10,7 +10,7 @@ const { URL } = require("whatwg-url");
 const MIMEType = require("whatwg-mimetype");
 const idlUtils = require("./jsdom/living/generated/utils.js");
 const VirtualConsole = require("./jsdom/virtual-console.js");
-const { createWindow } = require("./jsdom/browser/Window.js");
+const { createWindow } = require("./jsdom/browser/Window-impl.js");
 const { parseIntoDocument } = require("./jsdom/browser/parser");
 const { fragmentSerialization } = require("./jsdom/living/domparsing/serialization.js");
 const ResourceLoader = require("./jsdom/browser/resources/resource-loader.js");
