+++ lib/jsdom/browser/Window-impl.js
