+++ lib/jsdom/living/nodes/HTMLFrameElement-impl.js
@@ -5,7 +5,7 @@ const whatwgEncoding = require("whatwg-encoding");
 const { parseURL, serializeURL } = require("whatwg-url");
 const sniffHTMLEncoding = require("html-encoding-sniffer");
 
-const window = require("../../browser/Window");
+const window = require("../../browser/Window-impl.js");
 const HTMLElementImpl = require("./HTMLElement-impl").implementation;
 const { evaluateJavaScriptURL } = require("../window/navigation");
 const { reflectURLAttribute } = require("../../utils");
